src/pyproject.toml
[tool.poetry]
name = "video-app"
version = "0.1.0"
description = ""
authors = ["AntonZn"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
fastapi-versioning = "^0.5.0"
aerich = {extras = ["asyncpg"], version = "^0.5.0"}
aiofiles = "^0.6.0"
uvicorn = "^0.13.3"
gunicorn = "^20.0.4"
python-dotenv = "^0.15.0"
tortoise-orm = "0.16.21"
asyncpg = "^0.21.0"
fastapi-pagination = {extras = ["all"], version = "^0.5.2"}
aioboto3 = "^8.3.0"
PyJWT = "^2.1.0"
cryptography = "^3.4.7"
bcrypt = "^3.2.0"
pydantic = {version = "1.7.4", extras = ["email"]}
python-multipart = "^0.0.5"
filetype = "^1.0.7"
passlib = "^1.7.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
